+++ clang/lib/Sema/Scope.cpp
@@ -17,7 +17,7 @@
 
 using namespace clang;
 
-void Scope::setFlags(Scope *parent, unsigned flags) {
+void Scope::setFlags(Scope *parent, unsigned long flags) {
   AnyParent = parent;
   Flags = flags;
 
@@ -93,7 +93,7 @@ void Scope::setFlags(Scope *parent, unsigned flags) {
   }
 }
 
-void Scope::Init(Scope *parent, unsigned flags) {
+void Scope::Init(Scope *parent, unsigned long flags) {
   setFlags(parent, flags);
 
   DeclsInScope.clear();
@@ -113,7 +113,7 @@ bool Scope::containedInPrototypeScope() const {
   return false;
 }
 
-void Scope::AddFlags(unsigned FlagsToSet) {
+void Scope::AddFlags(unsigned long FlagsToSet) {
   assert((FlagsToSet & ~(BreakScope | ContinueScope | ContractAssertScope)) == 0 &&
          "Unsupported scope flags");
   if (FlagsToSet & BreakScope) {
@@ -198,13 +198,13 @@ void Scope::applyNRVO() {
 LLVM_DUMP_METHOD void Scope::dump() const { dumpImpl(llvm::errs()); }
 
 void Scope::dumpImpl(raw_ostream &OS) const {
-  unsigned Flags = getFlags();
+  unsigned long Flags = getFlags();
   bool HasFlags = Flags != 0;
 
   if (HasFlags)
     OS << "Flags: ";
 
-  std::pair<unsigned, const char *> FlagInfo[] = {
+  std::pair<unsigned long, const char *> FlagInfo[] = {
       {FnScope, "FnScope"},
       {BreakScope, "BreakScope"},
       {ContinueScope, "ContinueScope"},
